# src/config/swagger.yaml
openapi: "3.0.0"
security:
  - passwordAuth: []
info:
  title: "DockStatAPI"
  version: "2.0.1"
  externalDocs:
    description: DockStat(API) Wiki
    url: https://outline.itsnik.de/s/dockstat
  license:
    name: BSD-3-Clause
    url: https://github.com/Its4Nik/dockstatapi/tree/main?tab=BSD-3-Clause-1-ov-file#readme
  contact:
    email: info@itsnik.de
  description: |-
    ![DockStat](https://github.com/Its4Nik/dockstatapi/blob/dev/.github/DockStat-dark.png?raw=true)

    # Pipelines

    [![Docker Image CI](https://img.shields.io/github/actions/workflow/status/Its4Nik/dockstatapi/build-image.yml?branch=main&label=Docker%20Image%20CI&style=for-the-badge&logo=docker)](https://github.com/Its4Nik/dockstatapi/actions/workflows/build-image.yml)
    [![Validation](https://img.shields.io/github/actions/workflow/status/Its4Nik/dockstatapi/validation.yml?branch=dev&label=Validation&style=for-the-badge&logo=checkmarx)](https://github.com/Its4Nik/dockstatapi/actions/workflows/validation.yml)

    # Feature List:

    - Swagger API Documentation
    - Database (Keeps data for 24 hours max)
    - Advanced authentication using hashes and salt
    - `http` API to configure the backend
    - Multi-arch docker builds (using buildx github action)
    - Advanced security through middlewares: rate-limiting and authentication
    - Multi Arch Docker builds through docker buildx
    - High Availability using single master and unlimited worker nodes!

    # 🔗 DockStatAPI v2 Documentation

    _⚠️ = Deprecation warning_

...

...

...

...

...

...
servers:
...
paths:
...
  /frontend/show/{containerName}:
    post:
      tags:
        - "Frontend Configuration"
      operationId: frShowCon
      summary: Set `hide` to false for the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to unhide
      responses:
        "200":
          description: Success - now showing the container
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Container unhidden successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/hide/{containerName}:
    delete:
      tags:
        - "Frontend Configuration"
      operationId: frHideCon
      summary: Set `hide` to true for the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to unhide
      responses:
        "200":
          description: Success - now hiding the container
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Hid container succesfully"
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/tag/{containerName}/{tag}:
    post:
      tags:
        - "Frontend Configuration"
      operationId: frTagCon
      summary: Add a tag to the tag array for the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to add a tag to
        - name: tag
          in: path
          schema:
            type: string
          required: true
          description: The name of the tag to add
      responses:
        "200":
          description: Success - Tag added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Tag added successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/remove-tag/{containerName}/{tag}:
    delete:
      tags:
        - "Frontend Configuration"
      operationId: frRmTagCon
      summary: Remove the specified tag from the tag array for the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to remove a tag from
        - name: tag
          in: path
          schema:
            type: string
          required: true
          description: The name of the tag to remove
      responses:
        "200":
          description: Success - Tag removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Tag removed successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/pin/{containerName}:
    post:
      tags:
        - "Frontend Configuration"
      operationId: frPinCon
      summary: Set `pinned` to true for the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to pin
      responses:
        "200":
          description: Success - Container pinned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Container pinned successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/unpin/{containerName}:
    delete:
      tags:
        - "Frontend Configuration"
      operationId: frRmPinCon
      summary: Set `pinned` to false for the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to unpin
      responses:
        "200":
          description: Success - Container unpinned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Container unpinned successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/add-link/{containerName}/{link}:
    post:
      tags:
        - "Frontend Configuration"
      operationId: frAddLinkCon
      summary: Add a link to the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to add a link to
        - name: link
          in: path
          schema:
            type: URI
          required: true
          llowReserved: false
          description: The URI of the link (please use Uniform Resource Identifier format)
      responses:
        "200":
          description: Success - Link added to container successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Link added successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/remove-link/{containerName}:
    delete:
      tags:
        - "Frontend Configuration"
      operationId: frRmLinkCon
      summary: Remove a link to the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to remove a link from
      responses:
        "200":
          description: Success - Link removed from container successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Link removed successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/add-icon/{containerName}/{icon}/{useCustomIcon}:
    post:
      tags:
        - "Frontend Configuration"
      operationId: frAddIcon
      summary: Add an icon (path) to the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to add an icon to
        - name: icon
          in: path
          schema:
            type: string
          required: true
          description: The name of the icon file
        - name: useCustomIcon
          in: path
          schema:
            type: boolean
          required: false
          description: If the icon is a custom icon or not
      responses:
        "200":
          description: Success - Icon added to container successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Icon added successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
  /frontend/remove-icon/{containerName}:
    delete:
      tags:
        - "Frontend Configuration"
      operationId: frRmIcon
      summary: Removean icon from the specified container
      parameters:
        - name: containerName
          in: path
          schema:
            type: string
          required: true
          description: The name of the container to remove an icon from
      responses:
        "200":
          description: Success - Icon removed from container successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Icon removed successfully."
        "402":
          description: Error - Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/402"
        "403":
          description: Error - Password is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/403"
        "500":
          description: Error - Critical Error, please see the server's logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/500"
        "503":
          description: Error - The high-availability lock is currently active, please try again later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/503"
# ------------------------------
components:
  securitySchemes:
    passwordAuth:
      type: apiKey
      in: header
      name: x-password
      description: Password required for authentication
  schemas:
...
    402:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid password"
    403:
      type: object
      properties:
        status:
          type: string
          example: "denied"
        message:
          type: string
          example: "Password required"
    500:
      type: object
      properties:
        status:
          type: string
          example: "critical"
        message:
          type: string
          example: "Please see the server logs for more info"
    503:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Service unavailable. The high-availability lock is currently active. Please try again later."